+++ src/crons/cache.warmer/cache.warmer.service.ts
@@ -19,10 +19,10 @@ import { EsdtService } from "src/endpoints/esdt/esdt.service";
 import { CacheInfo } from "src/common/caching/entities/cache.info";
 import { TokenAssetService } from "src/endpoints/tokens/token.asset.service";
 import { GatewayComponentRequest } from "src/common/gateway/entities/gateway.component.request";
-import { MexSettingsService } from "src/endpoints/mex.analytics/mex.settings.service";
-import { MexEconomicsService } from "src/endpoints/mex.analytics/mex.economics.service";
-import { MexPairsService } from "src/endpoints/mex.analytics/mex.pairs.service";
-import { MexTokenService } from "src/endpoints/mex.analytics/mex.token.service";
+import { MexSettingsService } from "src/endpoints/mex/mex.settings.service";
+import { MexEconomicsService } from "src/endpoints/mex/mex.economics.service";
+import { MexPairsService } from "src/endpoints/mex/mex.pairs.service";
+import { MexTokenService } from "src/endpoints/mex/mex.token.service";
 
 @Injectable()
 export class CacheWarmerService {
