+++ src/endpoints/endpoints.controllers.module.ts
@@ -10,7 +10,7 @@ import { EndpointsServicesModule } from "./endpoints.services.module";
 import { HealthCheckController } from "./health-check/health.check.controller";
 import { IdentitiesController } from "./identities/identities.controller";
 import { KeysController } from "./keys/keys.controller";
-import { MexAnalyticsController } from "./mex.analytics/mex.analytics.controller";
+import { MexController } from "./mex/mex.controller";
 import { MiniBlockController } from "./miniblocks/mini.block.controller";
 import { NetworkController } from "./network/network.controller";
 import { NftController } from "./nfts/nft.controller";
@@ -41,7 +41,7 @@ import { WebsocketController } from "./websocket/websocket.controller";
     KeysController, MiniBlockController, NetworkController, NftController, TagController, NodeController,
     ProviderController, ProxyController, RoundController, SmartContractResultController, ShardController, StakeController, StakeController,
     TokenController, TransactionController, UsernameController, VmQueryController, WaitingListController,
-    HealthCheckController, DappConfigController, WebsocketController, MexAnalyticsController,
+    HealthCheckController, DappConfigController, WebsocketController, MexController,
   ],
 })
 export class EndpointsControllersModule { }
