+++ src/endpoints/mex/entities/mex.economics.ts
