+++ src/endpoints/mex/entities/mex.pair.ts
