+++ src/endpoints/mex/entities/mex.settings.ts
