+++ src/endpoints/transactions/transaction-action/recognizers/mex/mex.distribution.action.recognizer.service.ts
@@ -3,7 +3,7 @@ import { TransactionAction } from "../../entities/transaction.action";
 import { TransactionActionCategory } from "../../entities/transaction.action.category";
 import { TransactionMetadata } from "../../entities/transaction.metadata";
 import { MexFunction } from "./entities/mex.function.options";
-import { MexSettings } from "../../../../mex.analytics/entities/mex.settings";
+import { MexSettings } from "../../../../mex/entities/mex.settings";
 
 @Injectable()
 export class MexDistributionActionRecognizerService {
