+++ src/endpoints/transactions/transaction-action/recognizers/mex/transaction.action.mex.recognizer.module.ts
@@ -7,13 +7,13 @@ import { MexWrapActionRecognizerService } from "./mex.wrap.action.recognizer.ser
 import { MexDistributionActionRecognizerService } from "./mex.distribution.action.recognizer.service";
 import { TransactionActionModule } from "../../transaction.action.module";
 import { MexLockedAssetActionRecognizerService } from "./mex.locked.asset.action.recognizer.service";
-import { MexAnalyticsModule } from "src/endpoints/mex.analytics/mex.analytics.module";
+import { MexModule } from "src/endpoints/mex/mex.module";
 
 @Module({
   imports: [
     forwardRef(() => TokenModule),
     forwardRef(() => TransactionActionModule),
-    MexAnalyticsModule,
+    MexModule,
   ],
   providers: [
     TransactionActionMexRecognizerService,
